requirements.txt
twitchio
python-dotenv
requests
